[Chore] Enum, typealias 파일로 분리
